🖼️ Now sends a pre-defined chip image for sobriety of 365 days or less

+++ bot.py
@@ -16,12 +16,47 @@ class TweetStreamListener(tweepy.StreamListener):
     def on_status(self, tweet):
         logger.info(f"Tweeting to {tweet.user.name} ({tweet.user.screen_name})")
 
-        self.reply_to_tweet(tweet)
-        # self.like_tweet(tweet)
+        if tweet.in_reply_to_status_id is not None or tweet.user.id == self.me.id:
+            logger.info("Ignoring reply to or own tweet")
+            return
+
+        days = self.get_duration_in_days(tweet)
+        self.like_tweet(tweet)
+        self.reply_to_tweet(tweet, days)
 
     def on_error(self, status):
         logger.error("There was an error with the stream listener")
         return False
+    
+    def get_duration_in_days(self, tweet) -> int or None:
+        """
+        Determines if the tweet has a specified sobriety duration
+        Returns the length of sobriety in days
+
+        Returns:
+            int or None
+
+        """
+        regex = re.compile("([1-9][0-9]{0,2}) (days|weeks|months|years)", re.IGNORECASE)
+        match = regex.search(tweet.text)
+
+        if match:
+            number, duration = match.group().split(" ")
+
+            durations = {
+                "day": 1,
+                "days": 1,
+                "week": 7,
+                "weeks": 7,
+                "month": 30.4161,
+                "months": 30.4161,
+                "year": 365,
+                "years": 365
+            }
+
+            return round(int(number) * durations.get(duration, 1))
+        else:
+            return None
 
     def like_tweet(self, tweet):
         if not tweet.favorited:
@@ -30,31 +65,26 @@ class TweetStreamListener(tweepy.StreamListener):
             except Exception:
                 logger.error("Error favouriting tweet")
 
-    def reply_to_tweet(self, tweet):
-        # Ignore replies and own tweets
-        if tweet.in_reply_to_status_id is not None or tweet.user.id == self.me.id:
-            return
-
+    def reply_to_tweet(self, tweet, days=None):
         # TODO Remove - only for testing
         test_accounts = get_config()["TESTING_ACCOUNTS"]
         if tweet.user.screen_name not in test_accounts:
             return
+
+        media_ids = []
+        if days is not None:
+            chip = Chip(days)
+            media_obj = self.api.media_upload(chip.path)
+            media_ids.append(media_obj.media_id)
         
-        regex = re.compile("([1-9][0-9]{0,2}) (days|weeks)", re.IGNORECASE)
-        match = regex.search(tweet.text)
-        number, duration = match.group().split(" ")
-
-        chip = Chip(number)
-
-        media_obj = self.api.media_upload(chip.path)
-
         self.api.update_status(
-            status=f"@{tweet.user.screen_name} Congratulations!",
-            media_ids=[media_obj.media_id],
+            status=f"Congratulations @{tweet.user.screen_name}!",
+            media_ids=media_ids,
             in_reply_to_status_id=tweet.id
         )
 
-        chip.delete()
+        if days is not None:
+            chip.delete()
 
 
 def main():
